+++ src/main/java/com/example/bankcards/security/UserPrincipal.java
@@ -2,19 +2,19 @@ package com.example.bankcards.security;
 
 public class UserPrincipal {
 
-    private final int userId;
+    private final Long userId;
 
     private final String username;
 
     private final String role;
 
-    public UserPrincipal(int userId, String username, String role) {
+    public UserPrincipal(Long userId, String username, String role) {
         this.userId = userId;
         this.username = username;
         this.role = role;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
